requirements.txt
playwright
playwright-stealth
beautifulsoup4
lxml
typer
rich
sqlalchemy
psycopg2-binary
python-dotenv
apscheduler
